Makes CharacterNameForm prop-free
